backend/requirements.txt
alembic==1.15.2
bcrypt==4.0.1
email-validator==2.2.0
fastapi==0.115.12
httpx==0.28.1
openai==1.76.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-docx==1.1.2
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
pypdf==5.4.0
qdrant-client==1.13.3
sqlalchemy==2.0.40
uvicorn[standard]==0.34.2
